+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Upgrade dependencies
 + Add level to getBackgroundColor
 + Add level to getColor
++ Refactor get message
 
 ## 1.0.0
 + Upgrade dependencies